chore: move unauthorized to the components folder
